update public Rust API
The gist of it is that we now are dealing with `SourceEntry` instead of
`GlobEntry`. A `SourceEntry` in the `node` crate maps to a
`PublicSourceEntry`. This makes it easy for JS APIs to pass in the
values.

A `PublicSourceEntry` looks like:

- `base` — the base path of the source
- `pattern` — the glob pattern
- `negated` — whether or not the pattern should be negated

Internally we will map this to a `SourceEntry`, this is a proper enum
that looks like this:
```rs
pub enum SourceEntry {
    Auto { base: PathBuf },
    IgnoredAuto { base: PathBuf },
    Pattern { base: PathBuf, pattern: String },
    IgnoredPattern { base: PathBuf, pattern: String },
}
```

Before we construct a Scanner, we will also make sure to optimize these
patterns. Some optimization steps:

1. Each `pattern` will be brace-expanded and a new `SourceEntry` will be
   created. This allows us to always deal with simple patterns.
2. The `base` of each `SourceEntry` will be canonicalized so we are
   dealing with the real paths and symlinks are resolved.
3. When patterns include static parts such as `/src/*.html`, then the
   static part (in this case `src`) will be moved to the `base`.